test(editor): share the editor mount helpers
This PR added a second copy of mount() and flush(), line for line the
same as the ones in attachment.flow.test.ts. The new copy was the typed
one, so the repo was left with a typed and an untyped version side by
side. Both now come from src/molecules/editor/test-helpers.ts, which
stays extension-agnostic: each test passes its own kit.

The helper is not public API — index.ts does not re-export it, there is
no lib build entry, and propsgen only walks component folders, so the
API tables are unchanged.
