+++ src/molecules/editor/extensions/attachment/attachment.flow.test.ts
@@ -5,50 +5,23 @@
  * upload function, then `uploadAttachmentFiles` (what media-drop / EditorDropZone
  * call). Asserts a chip node is actually inserted and the NodeView mounts.
  */
-import { describe, it, expect, beforeEach, vi } from 'vitest'
-import { createApp, h, nextTick, reactive } from 'vue'
-
-let Editor: any
-let EditorContent: any
-let CommentKit: any
-
-beforeEach(async () => {
-  ;({ default: Editor } = await import('../../Editor.vue'))
-  ;({ default: EditorContent } = await import('../../EditorContent.vue'))
-  ;({ CommentKit } = await import('../../kits'))
-})
-
-function mount(staticProps: Record<string, any>) {
-  const state = reactive<Record<string, any>>({ modelValue: '' })
-  let editor: any = null
-  const root = document.createElement('div')
-  document.body.appendChild(root)
-  const app = createApp({
-    render() {
-      return h(Editor, { ...staticProps, ...state }, {
-        default: ({ editor: e }: any) => {
-          editor = e
-          return h(EditorContent, { editor: e })
-        },
-      })
-    },
-  })
-  app.mount(root)
-  return { getEditor: () => editor, app, root }
-}
-
-const flush = async () => {
-  for (let i = 0; i < 5; i++) {
-    await Promise.resolve()
-    await nextTick()
-  }
-}
+import { describe, it, expect, vi } from 'vitest'
+import { CommentKit } from '../../kits'
+import { flush, mount } from '../../test-helpers'
+import type { Editor as TiptapEditor } from '../../useEditor'
 
 describe('attachment drop flow (CommentKit)', () => {
   it('inserts a chip node when a non-media file is uploaded', async () => {
-    const upload = vi.fn(async (file: File) => ({ file_url: `/files/${file.name}` }))
+    const upload = vi.fn(async (file: File) => ({
+      file_url: `/files/${file.name}`,
+    }))
     const ctx = mount({ extensions: [CommentKit], uploadFunction: upload })
-    const editor = ctx.getEditor()
+    // tiptap's `Storage` interface is empty and nothing here augments it, so
+    // the storage diagnostic below reads through a widened handle. Same cast
+    // `resolveUploadOptions` makes.
+    const editor = ctx.getEditor() as TiptapEditor & {
+      storage: Record<string, any>
+    }
 
     // diagnostics: does the editor expose the command + upload storage?
     expect(typeof editor.commands.uploadAttachmentFiles).toBe('function')
